feat(flags): Include --ignore flag for regex folder paths and rework --hidden-files flag
