+++ src/rejx/__init__.py
@@ -1,24 +1,15 @@
 from rejx.cli import (
     app,
-    apply_changes,
-    build_file_tree,
     clean,
     diff,
-    find_rej_files,
     fix,
     ls,
-    parse_rej_file,
-    process_rej_file,
 )
 
 __all__ = [
-    "apply_changes",
-    "build_file_tree",
+    "app",
     "clean",
     "diff",
-    "find_rej_files",
     "fix",
     "ls",
-    "parse_rej_file",
-    "process_rej_file",
 ]
