+++ src/rejx/logger.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+import logging
+
+from rich.logging import RichHandler
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(message)s",
+    datefmt="[%X]",
+    handlers=[RichHandler(rich_tracebacks=True)],
+)
+
+console = logging.StreamHandler()
+console.setLevel(logging.DEBUG)
+
+formatter = logging.Formatter("%(name)-12s: %(levelname)-8s %(message)s'")
+console.setFormatter(formatter)
+logging.getLogger("").addHandler(console)
+
+logger = logging.getLogger(__name__)
