+++ tests/data/.hidden-sample.txt
@@ -1,0 +1,3 @@
+Line 1
+Line 2 - Modified
+Line 3
